heal7-project/backend/services/paperwork-service/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
pyyaml>=6.0.0
aiohttp>=3.8.0
pytesseract>=0.3.10
Pillow>=10.0.0
openai>=1.0.0